Synchronizing with Portage. Adding sigprune_fet format. Fixing the UNK token which could have been part of the corpus vocabulary. UNK_TOKEN = " UNK " which make it impossible to have in the corpus's vocabulary.

+++ src/tpt/textpt2tppt.sh
@@ -111,7 +111,7 @@ for x in tppt cbk trg.repos.dat src.tdx trg.tdx; do
       error_exit "Can't mv $OUTPUTPT.$x into $OUTPUTPT$TPT_EXTENSION/$x, model probably exists but can't be moved or renamed properly."
 done
 cd ..
-rm -r $TMPDIR
+[[ ! $DEBUG ]] && rm -r $TMPDIR
 
 echo "
 The five files tppt, cbk, src.tdx, trg.tdx and trg.repos.dat, together,
